Add hooks to handle the position of filter
What solved
[X] Add hook to handle the position of the filter
[X] Add the hook to reserve chart and line chart
