+++ src/views/Finances/FinancesView.tsx
@@ -125,7 +125,7 @@ const FinancesView: React.FC<Props> = ({ budgets, allBudgets, yearsRange, initia
           <ContainerSpace>
             <ExpenseMetricsSection
               expenseMetrics={{
-                title: levelNumber === 1 ? 'Sky Expense Metrics' : 'Expense Metrics',
+                title: levelNumber === 1 ? 'Sky Expense Metrics' : title,
                 selectedGranularity: expensesMetrics.selectedGranularity,
                 isCumulative: expensesMetrics.isCumulative,
                 cumulativeType: expensesMetrics.cumulativeType,
